requirements.txt
agno
google-genai
python-dotenv
markitdown[all]
pyyaml
streamlit
lancedb
pandas
langfuse
openlit
fastapi